function context fixed: Function-Context is now set to a special null-pattern to prevent conversion-errors.
Lookahead-error fixed: The scanner was reading one char to much (i.e. i
=) thereby causing errors in the symbol table. This error was caused by
the line identifier {...=yytext} and was replaced by identifier
{...=strdpl(yytext)}

Also the header tokentype.h was added to the library to provide an
unified header for all tokens. The header itself was not implemented
into the parser yet.

There still persists an error where the parameter-header is null during
function-declaration.